+++ src/filters/Noise.ts
@@ -2,7 +2,7 @@ import type { TClassProperties } from '../typedefs';
 import { BaseFilter } from './BaseFilter';
 import type { T2DPipelineState, TWebGLUniformLocationMap } from './typedefs';
 import { classRegistry } from '../ClassRegistry';
-import { fragmentSource } from './shaders/noise.shaders';
+import { fragmentSource } from './shaders/noise';
 
 export const noiseDefaultValues: Partial<TClassProperties<Noise>> = {
   mainParameter: 'noise',
